[BUGFIX] Remove existing Solr document id
Apache Solr for TYPO3 requires a special document id structure
that is build within the SiteHashIndexingFilter.

As in newer Nutch versions an id is already prepared and to
prevent that a multiple field is used for the id, an existing
id field is now removed before the custom id is added.